fix(CardHeader): remove styled-component
