+++ src/components/Analytics/Analytics.tsx
@@ -4,7 +4,7 @@ import styled from 'styled-components';
 import { useAllowedHireStreams } from '../../modules/hireStreamsHooks';
 import { Paths } from '../../utils/paths';
 import { LayoutMain } from '../LayoutMain/LayoutMain';
-import { CardHeader } from '../CardHeader';
+import { CardHeader } from '../CardHeader/CardHeader';
 import { QueryResolver } from '../QueryResolver/QueryResolver';
 import { HireStreamAnalyticsList } from '../HireStreamAnalyticsList';
 import { Card } from '../Card';
