+++ src/components/TitleMenu/TitleMenu.module.css
@@ -1,5 +1,4 @@
 .TitleMenu {
    display: inline-block;
-   padding-left: var(--gap-xs);
 }
